% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_ubc_cv.R
\name{theme_ubc_cv}
\alias{theme_ubc_cv}
\title{Theme for UBC CV tables}
\usage{
theme_ubc_cv(x)
}
\arguments{
\item{x}{flextable object}
}
\value{
flextable object
}
\description{
Theme for UBC CV tables
}
\examples{
library(flextable)
library(ubccv)

flextable(mtcars) \%>\%
 theme_ubc_cv()
}
